refactor : 투표 페이지 투표 결과 조회 컴포넌트 분리 #4
